pyproject.toml
[project]
name = "manor-lords-ai-guide"
version = "0.1.0"
description = "AI Strategic Guide System for Manor Lords"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sse-starlette>=2.3",
    "pydantic>=2.11",
    "pyyaml>=6.0",
    "watchdog>=6.0",
    "google-genai>=1.14",
]
